+++ src/lib.ts
@@ -1,6 +1,6 @@
 import * as vscode from "vscode";
 import { Message, nextFuzzyMessage, nextUntranslatedMessage } from "./message";
-import { moveCursorTo } from "./moving";
+import { focusOnNextTarget } from "./focusing";
 
 export function moveToNextUntranslatedMessage(editor: vscode.TextEditor) {
   focusOnNextTarget(editor, nextUntranslatedMessage);
@@ -16,23 +16,4 @@ export function moveToNextFuzzyMessage(editor: vscode.TextEditor) {
 
 export function moveToPreviousFuzzyMessage(editor: vscode.TextEditor) {
   focusOnNextTarget(editor, nextFuzzyMessage, true);
-}
-
-function focusOnMessage(editor: vscode.TextEditor, message: Message) {
-  const position = moveCursorTo(editor, message.msgstrLine, 8);
-  editor.revealRange(
-    new vscode.Range(position, position),
-    vscode.TextEditorRevealType.InCenterIfOutsideViewport
-  );
-}
-function focusOnNextTarget(
-  editor: vscode.TextEditor,
-  nextTargetFunc: Function,
-  backwards = false
-) {
-  const position = editor.selection.active;
-  const message = nextTargetFunc(editor.document, position.line, backwards);
-  if (message !== null) {
-    focusOnMessage(editor, message);
-  }
 }
